automated_blog_content_pipeline/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
openai>=1.12.0
faiss-cpu>=1.7.4
tiktoken>=0.5.2
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
pypdf>=3.17.0